added logic and routes for deleting a show

+++ controllers/controllers.js
@@ -107,9 +107,10 @@ exports.getAllShowsByUser = (req, res) => {
 exports.deleteShowByUser = (req, res) => {
   knex('mylist')
       .delete()
-      .where({showId: req.params.showId})
+      .where({showId: req.params.showId, users_id: req.params.userId})
       .then(() => {
           res.status(204).send(`User has successfully deleted show with show ID: ${req.params.showId} `)
+          
       })
       .catch(err => {
           res.status(400).send(`Error deleting show with showID: ${req.params.showId}`)
